+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Documentation can be found at [ReadTheDocs](https://dynamic-surroundings.readthe
 **Fixes**
 * Configure sound button on Sound Options screen will now work when Music Manager replacement is toggled off.
 
+**Changes**
+* Consolidated the Block and Cull buttons in the Individual Sound configuration menu into a single tri-state button. The icons didn't grow on me.
+
 > ### DynamicSurroundings-1.21.1-0.4.4
 
 Documentation can be found at [ReadTheDocs](https://dynamic-surroundings.readthedocs.io/en/latest/index.html) website.